Fix recipes not being able to be created
I originally used multimodelform v3.0, which had a bug where it would
show table headers even for hidden form elements, so I simply edited
the extension's code to just show all fields. This inadvertently
allowed the id column to be a hidden field with a blank value. This
actually worked until the following yii commit which fixed the bug
allowing this.

https://github.com/yiisoft/yii/commit/c1538e5d047dcbb163fb591021380250d4f609da

multimodelform v3.1 fixed the bug I was working around, so after
upgrading it, I reverted back to the right way, and the world is back
to normal again! :)